# customer-app/web-customer/docs/openapi.yaml
openapi: 3.0.0
info:
  title: FlashFit API
  version: 1.0.0
  description: API documentation for FlashFit Customer App
servers:
  - url: https://flashfit.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local server
paths:
  /products:
    get:
      summary: Get products
      parameters:
        - in: query
          name: pincode
          schema:
            type: string
          required: true
          description: Filter products by serviceable pincode
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /createOrder:
    post:
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '200':
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId:
                    type: string
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        price:
          type: number
        category:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
    OrderRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        totalAmount:
          type: number
    CartItem:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
        price:
          type: number
    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        pincode:
          type: string